feat: Enhance getAllPosts method to support pagination and filtering, update related services and DTOs

+++ src/main/java/org/amalitech/bloggingplatformspring/dtos/responses/PageResponse.java
@@ -7,6 +7,6 @@ public record PageResponse<T>(
         int page,
         int size,
         String sort,
-        int totalElements
+        long totalElements
 ) {
 }
